fix : metadata 위치 수정
